@@ -47,35 +47,35 @@
     "date-fns": "2.0.0-alpha.1",
     "element-closest": "^2.0.2",
     "is-ua-webview": "^1.0.0",
-    "popper.js": "^1.14.0"
+    "popper.js": "^1.14.0",
+    "tslib": "^1.10.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.0",
-    "@angular-devkit/build-ng-packagr": "~0.7.0",
-    "@angular/animations": "^6.1.0",
-    "@angular/cli": "~6.1.1",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
-    "@edcarroll/code-style": "^2.0.0",
-    "@types/extend": "^3.0.0",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "@types/prismjs": "~1.4.18",
+    "@angular-devkit/build-angular": "~0.901.15",
+    "@angular-devkit/build-ng-packagr": "~0.901.15",
+    "@angular/animations": "^9.1.13",
+    "@angular/cli": "~9.1.15",
+    "@angular/common": "^9.1.13",
+    "@angular/compiler": "^9.1.13",
+    "@angular/compiler-cli": "^9.1.13",
+    "@angular/core": "^9.1.13",
+    "@angular/forms": "^9.1.13",
+    "@angular/language-service": "^9.1.13",
+    "@angular/platform-browser": "^9.1.13",
+    "@angular/platform-browser-dynamic": "^9.1.13",
+    "@angular/router": "^9.1.13",
+    "@types/extend": "3.0.0",
+    "@types/jasmine": "2.8.6",
+    "@types/jasminewd2": "2.0.3",
+    "@types/node": "^12.11.1",
+    "@types/prismjs": "1.4.18",
     "bowser": "^1.7.2",
+    "codelyzer": "^5.1.2",
     "core-js": "^2.5.4",
     "date-fns": "2.0.0-alpha.1",
     "element-closest": "^2.0.2",
     "extend": "^3.0.1",
-    "fs-extra": "^5.0.0",
+    "fs-extra": "5.0.0",
     "is-ua-webview": "^1.0.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
@@ -87,16 +87,14 @@
     "karma-min-reporter": "^0.1.0",
     "karma-phantomjs-launcher": "^1.0.4",
     "karma-remap-istanbul": "~0.4.0",
-    "ng-packagr": "^3.0.0",
+    "ng-packagr": "^9.0.0",
     "popper.js": "^1.14.0",
     "prismjs": "~1.6.0",
-    "rxjs": "^6.0.0",
+    "rxjs": "6.6.7",
     "ts-node": "~5.0.1",
-    "tsickle": ">=0.29.0",
-    "tslib": "^1.9.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "~0.8.26"
+    "typescript": "3.8.3",
+    "web-animations-js": "^2.3.2",
+    "zone.js": "~0.10.2"
   }
 }
