@@ -1,6 +1,6 @@
 {
-  "name": "ng2-semantic-ui",
-  "version": "0.10.0-alpha.6",
+  "name": "ng2-semantic-ui-v9",
+  "version": "0.0.1",
   "description": "Angular Semantic UI Components",
   "repository": {
     "type": "git",
@@ -28,10 +28,10 @@
     "popper.js": "^1.14.0"
   },
   "peerDependencies": {
-    "@angular/common": ">= 6.0.0",
-    "@angular/compiler": ">= 6.0.0",
-    "@angular/core": ">= 6.0.0",
-    "@angular/forms": ">= 6.0.0",
-    "rxjs": "^6.0.0"
+    "@angular/common": ">= 9.1.13",
+    "@angular/compiler": ">= 9.1.13",
+    "@angular/core": ">= 9.1.13",
+    "@angular/forms": ">= 9.1.13",
+    "rxjs": "6.6.7"
   }
 }
